LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-with-step-expected.txt

Test if step limits available dates in calendar picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Step without min.
PASS availableDayCells() is "2011-05-08,2011-05-17,2011-05-26,2011-06-04"
Step when min is set.
PASS availableDayCells() is "2011-05-01,2011-05-10,2011-05-19,2011-05-28"
PASS successfullyParsed is true

TEST COMPLETE
 
